Updating documentation to reflect changes to the split transactions API (#15)
* Updating documentation to reflect changes to the split transactions API.

We no longer accept a separate note within a split transaction, in order to more closely match behaviour in the current iteration of the web app.

Likewise, an is_transfer parameter is now accepted for transaction splits, to indicate that the split should be treated as a transfer.

Additional information and an updated example for the transaction amount parameter is also included, clarifying that transaction type is indicated via the sign provided on the transaction amount.

* Further updates to negative based examples

* Improving description of the signs on transaction amounts, updating more debit transaction examples to use a signed amount.

* Further tweaks, clarifications and unifications.

---------

Co-authored-by: James Wigglesworth <james+macbook-pro-2021@pocketsmith.com>